회원가입 API 구현

+++ src/main/java/com/reday/global/security/UserPrincipal.java
@@ -16,7 +16,7 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class UserPrincipal implements UserDetails {
 
-	private final Long memberIdx;
+	private final Integer memberIdx;
 	private final String email;
 	private final String password;
 	private final Collection<? extends GrantedAuthority> authorities;
